+++ apps/web/src/components/mainpage/coreValue/CoreValueSection.tsx
@@ -1,7 +1,11 @@
-import clsx from 'clsx';
-import { FC, ReactNode, useEffect, useRef, useState } from 'react';
+'use client';
 
-import HorizontalScroll from './HorizontalScroll';
+import clsx from 'clsx';
+import { FC } from 'react';
+
+import HorizontalScroll from '@/components/common/HorizontalScroll';
+import InspectableBox from '@/components/common/InspectableBox';
+
 import { BaseImage, ConnectionImage, OpportunityImage, OwnershipImage, PleasureImage } from './images';
 import ValueCard from './ValueCard';
 
@@ -18,19 +22,19 @@ const CoreValueSection: FC<CoreValueSectionProps> = ({}) => {
               <span>VALUE</span>
             </div>
             {cards.map((card, idx) => (
-              <Trackable key={card.keyword}>
+              <InspectableBox key={card.keyword}>
                 {({ x, width }) => (
                   <ValueCard
                     className='h-[52rem] w-[36.8rem]'
                     {...card}
                     seq={idx + 1}
                     centerLineProgress={centerLine}
-                    center={x + width / 2}
+                    flipRange={[x - width / 2, x + width / 2]}
                   />
                 )}
-              </Trackable>
+              </InspectableBox>
             ))}
-            <div className='w-[70vw]' />
+            <div className='w-[50vw]' />
           </div>
         </div>
       )}
@@ -38,67 +42,7 @@ const CoreValueSection: FC<CoreValueSectionProps> = ({}) => {
   );
 };
 
-interface TrackableProps {
-  className?: string;
-  children: (ctx: PositionContext) => ReactNode;
-}
-
-interface PositionContext {
-  width: number;
-  height: number;
-  x: number;
-  y: number;
-}
-
-const Trackable: FC<TrackableProps> = ({ className, children }) => {
-  const ref = useRef<HTMLDivElement>(null);
-
-  const [positionContext, setPositionContext] = useState<PositionContext | null>(null);
-
-  useEffect(() => {
-    if (!ref.current) {
-      return;
-    }
-
-    const observer = new ResizeObserver((entires) => {
-      if (!ref.current) {
-        return;
-      }
-
-      for (const entry of entires) {
-        setPositionContext({
-          height: entry.contentRect.height,
-          width: entry.contentRect.width,
-          x: ref.current.offsetLeft,
-          y: ref.current.offsetTop,
-        });
-      }
-    });
-
-    observer.observe(ref.current);
-
-    return () => observer.disconnect();
-  }, []);
-
-  return (
-    <div ref={ref} className={className}>
-      {positionContext && children(positionContext)}
-    </div>
-  );
-};
-
-interface CardBaseProps {
-  className?: string;
-  children?: ReactNode;
-}
-
-const DescriptionCard: FC<CardBaseProps> = ({ className, children }) => {
-  return (
-    <div className={clsx('flex h-full w-full items-center justify-center whitespace-pre-line text-center', className)}>
-      {children}
-    </div>
-  );
-};
+const descriptionStyles = 'flex h-full w-full items-center justify-center whitespace-pre-line text-center';
 
 const cards = [
   {
@@ -106,12 +50,12 @@ const cards = [
     name: '오너십',
     image: <OwnershipImage />,
     description: (
-      <DescriptionCard className='bg-brand-orange text-orange-sub'>
+      <div className={clsx(descriptionStyles, 'bg-brand-orange text-orange-sub')}>
         {`주체적인 책임감으로
           스스로 필요한 가치를 발견하고,
           모두가 앞으로 나아가는
           추진력을 경험합니다.`}
-      </DescriptionCard>
+      </div>
     ),
   },
   {
@@ -119,11 +63,11 @@ const cards = [
     name: '연결',
     image: <ConnectionImage />,
     description: (
-      <DescriptionCard className='bg-brand-skyblue text-skyblue-sub'>
+      <div className={clsx(descriptionStyles, 'bg-brand-skyblue text-skyblue-sub')}>
         {`자유로운 네트워킹 환경 속
           뛰어난 동료와 깊이 연결되는
           경험을 할 수 있습니다.`}
-      </DescriptionCard>
+      </div>
     ),
   },
   {
@@ -131,11 +75,11 @@ const cards = [
     name: '기반',
     image: <BaseImage />,
     description: (
-      <DescriptionCard className='bg-brand-blue text-blue-sub'>
+      <div className={clsx(descriptionStyles, 'bg-brand-blue text-blue-sub')}>
         {`조직의 단단한 기반으로써,
           성장에 필요한 믿음직한 발판을
           제공합니다.`}
-      </DescriptionCard>
+      </div>
     ),
   },
   {
@@ -143,12 +87,12 @@ const cards = [
     name: '기회',
     image: <OpportunityImage />,
     description: (
-      <DescriptionCard className='bg-brand-pink text-pink-sub'>
+      <div className={clsx(descriptionStyles, 'bg-brand-pink text-pink-sub')}>
         {`지속적인 경험을 통해
           시야를 확장하고,
           더 넓고 새로운 경험의
           기회를 선사합니다.`}
-      </DescriptionCard>
+      </div>
     ),
   },
   {
@@ -156,12 +100,12 @@ const cards = [
     name: '즐거움',
     image: <PleasureImage />,
     description: (
-      <DescriptionCard className='bg-brand-yellow text-yellow-sub'>
+      <div className={clsx(descriptionStyles, 'bg-brand-yellow text-yellow-sub')}>
         {`존중을 기반으로,
           진정으로 모두가 즐겁게
           스스로의 목표를 세우고
           이루는 경험을 선사합니다.`}
-      </DescriptionCard>
+      </div>
     ),
   },
 ];
